Renames package to pyweconnect
