+++ pyweconnect/const.py
