+++ pyweconnect/utilities.py
