+++ pyweconnect/vehicle.py
@@ -8,7 +8,7 @@ from collections import OrderedDict
 from datetime import datetime, timedelta, timezone
 from json import dumps as to_json
 
-from volkswagencarnet.vw_utilities import find_path, is_valid_path
+from pyweconnect.utilities import find_path, is_valid_path
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -563,7 +563,7 @@ class Vehicle:
 
     def dashboard(self, **config):
         # Classic python notation
-        from volkswagencarnet.vw_dashboard import Dashboard
+        from pyweconnect.dashboard import Dashboard
         return Dashboard(self, **config)
 
     @property
