+++ setup.py
@@ -1,4 +1,4 @@
-from setuptools import setup
+from setuptools import setup, find_packages
 
 # read the contents of your README file
 from os import path
@@ -7,12 +7,13 @@ with open(path.join(this_directory, 'README.md'), encoding='utf-8') as f:
     long_description = f.read()
 
 setup(
-    name='volkswagencarnet',
+    name='pypyweconnect',
     description='Communicate with Volkswagen WeConnect',
     author='Robin Ostlund',
     author_email='me@robinostlund.name',
     url='https://github.com/robinostlund/volkswagencarnet',
     long_description=long_description,
+    packages=find_packages(),
     long_description_content_type='text/markdown',
     install_requires=list(open("requirements.txt").read().strip().split("\n")),
     use_scm_version=True,
