+++ tests/dummy_test.py
@@ -11,7 +11,7 @@ sys.path.insert(0, myPath + '/../')
 
 @pytest.mark.asyncio
 async def test_volkswagencarnet():
-    from volkswagencarnet.vw_connection import Connection
+    from pyweconnect.connection import Connection
     async with ClientSession() as session:
         connection = Connection(session, 'test@example.com', 'test_password')
         # if await connection._login():
